Seperate out complex state object to multiple
This is to fix https://github.com/wso2-enterprise/choreo/issues/12356
after a refactoring done for Expression Editor. Having all the states
in a single complex state object was causing issues when setState calls
are batched.

With the new changes to expression editor, everytime the user changes
the content, it will invoke both validate and onUpdate call backs. In
Listener form, both of these calls comes back to same state setter.

Hence react somehow batch process these state update requests and misses
some of the updates to state while doing the batch processing.

We need to investigate this further to understand where the root cause
is. However, as a rule of thumb, we need to avoid creating complex objects
which aggregates all the differrent states we need to keep in a component.