Add a choreo metadata file to project structure
This will be used to activate the extension now. Previously
we were depending on onStartupFinished event. But it occurs
only after vscode loading is complete - which can take >30secs
in some cases.

When we have a custom file in the file hierarchy, we can use
it to detect the choreo project and activate the extension very
early in the vscode startup.