Allow listing all files in Dockerfile dialog
VSCode API expects filters to be file extensions.
In docker's case, Dockerfile is the name of the
file. Not a file extension.

Hence when we have 'Dockerfile' as the filter, it fails
to list down dockerfiles.
Furthermore, there can be situations where people using
different suffixes, eg: Dockerfile.dev etc to have multiple
dockerfiles in the same location.