Hide any orgs with zero repositories configured
This is to hide the gh orgs which have zero configured
repositories. For free users, only public repos are allowed.
For some reason, if the user in free tier, they still get the
list of orgs which have private repos, but just the org name.

This is to filter out them