Change org fetch depending on token
If the user is signing in, we need to call validateUser
api to validate the user and retrieve the org list. But
if the user is reusing an existing token, we need to
switch to orgs API as validation does not need to be done.

This is to fix the intermittant logout issues during
loging from existing sessions where we previously called
validate user api in both cases.

Related https://github.com/wso2-enterprise/choreo/issues/21134