Use a seperate hook to manage selected org, etc
Previously, this was managed in ChoreoWebviewContext.
However, context also managed the login status, etc.
Whenever there are frequent parallel updates to both login
status and the selected orgs, some of the components
subscribed to context updates receives invalid status
for selected org.

This is to seperate those concerns & fix it.